roleeee

+++ src/main/java/rs/htec/cms/cms_bulima/constants/TableConstants.java
@@ -11,9 +11,9 @@ package rs.htec.cms.cms_bulima.constants;
  */
 public interface TableConstants {
     
-    public static final int CMS_USER = 1;
-    public static final int NEWS = 2;
-    public static final int SLIDER_CONTENT = 3;
-    public static final int QUESTION_OF_THE_DAY = 4;
-    public static final int QUESTION_OF_THE_DAY_PRIZE = 5;
+    public static final long CMS_USER = 1;
+    public static final long NEWS = 2;
+    public static final long SLIDER_CONTENT = 3;
+    public static final long QUESTION_OF_THE_DAY = 4;
+    public static final long QUESTION_OF_THE_DAY_PRIZE = 5;
 }
